fix: resolve CI failures on PR #28361
`check-typecheck-boundaries.mjs` pins an exact importer set for
`src/signals/route.ts`, so the new `legacy-zero-compatibility.test.ts`
failed `lint-types-apps` as a fifth importer.

Fold its assertions into the two files that already own this boundary
instead of widening the allowlist: the compatibility-table invariants go
to `api-namespace-compatibility.test.ts`, which already owns the
`withApiNamespaceAliases` route-table properties, and the provider-console
assertion becomes one check inside the existing per-path loop in
`provider-console-paths.test.ts`, reusing `FINAL_CONSOLE_ROUTES` rather
than restating those eight paths a third time.

Also correct the `LEGACY_ZERO_PATHS` comment: while the fallback is still
in place, removing a row makes a path start reporting rather than start
404ing, and reachability follows the table only after #26701.
